test(klog_daemon): split cluster integration tests by feature
